Moved OrderDetails to components/orders
